LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/replaced-elements/embedded-content/video-controls-stacking-hit-test-expected.txt

Button over video

FAIL video-controls-stacking-hit-test assert_equals: Hit test in the button area should hit the button expected Element node <button>Button over video</button> but got Element node <video controls=""></video>
